requirements.txt
Flask>=1.0.2,<=2
Pillow>=5.3.0,<=6
boto3>=1.9.70,<=2
ruamel.yaml==0.15.81
requests>=2.21.0,<=3
werkzeug
botocore